backend/admin-dashboard/requirements.txt
# Admin Dashboard API Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pyyaml>=6.0.1
httpx>=0.26.0

# Optional: for production
gunicorn>=21.2.0
